🔊 add logging to setup

+++ setup.sh
@@ -16,12 +16,16 @@ for item in "$VIRTUAL_HOME"/* "$VIRTUAL_HOME"/.[^.]*; do
 	fi
 done
 
-bun i -g cowsay
-ls -l "$HOME/.bun/install"
-rm -rf "$HOME/.bun/install/global"
-echo "Removing old global bun folder"
-ln -sf "$VIRTUAL_HOME"/.bun/install/global "$HOME/.bun/install/global"
-echo "Symlinked new global bun folder"
-ls -l "$HOME/.bun/install"
+# if there's no global bun folder, create one
+if [ ! -d "$HOME/.bun/install/global" ]; then
+    echo "Bun hasn't been set up yet, installing cowsay in order to create ~/.bun/install/global"
+    bun i -g cowsay
+    cowsay "Job's done! Now we'll remove the global folder and link it out from the repo."
+    rm -rf "$HOME/.bun/install/global"
+    ln -sf "$VIRTUAL_HOME"/.bun/install/global "$HOME/.bun/install/global"
+    cowsay "Symlinked new global bun folder"
+fi
+
+./scripts/npm-i.sh
 
 echo "All dotfiles and configurations have been symlinked."
